+++ bms-frontend/src/App.jsx
@@ -28,7 +28,7 @@ function App() {
   const isCheckoutPage = useMatch("/shows/:showId/:state/checkout");
 
   return (
-    <div className="flex flex-col min-h-screen">
+    <div className="flex flex-col min-h-screen bg-white dark:bg-gray-900 transition-colors duration-300">
       {/* Header */}
       {!isSeatLayoutPage && !isCheckoutPage && <Header />}